+++ variables.tf
@@ -25,6 +25,12 @@ variable "subnet_id" {
   type        = string
   description = "The ID of the Subnet which the Application Gateway should be connected to."
 }
+variable "waf_configuration" {
+  type        = map(string)
+  default     = {}
+  description = ""
+  # waf_configuration = { enabled = "", firewall_mode = "", rule_set_version = ""}
+}
 variable "frontend_ip_configuration" {
   type        = map(string)
   description = "A mapping the front ip configuration."